Split pr_cmd_inline publication finish tests into modules
