+++ adl/src/cli/tests/pr_cmd_inline/finish/publication.rs
@@ -1,1405 +1,6 @@
 use super::*;
-use crate::cli::pr_cmd::github::{
-    attach_post_merge_closeout, attach_pr_janitor, current_pr_url,
-    ensure_or_repair_pr_closing_linkage, ensure_pr_closing_linkage, pr_has_closing_linkage,
-};
 
-#[test]
-fn real_pr_finish_creates_draft_pr_and_commits_branch_changes() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-create");
-    let origin = temp.join("origin.git");
-    let repo = temp.join("repo");
-    fs::create_dir_all(&repo).expect("repo dir");
-    copy_bootstrap_support_files(&repo);
-    init_git_repo(&repo);
-    assert!(Command::new("git")
-        .args(["config", "user.name", "Test User"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    assert!(Command::new("git")
-        .args(["config", "user.email", "test@example.com"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    fs::write(repo.join(".gitignore"), ".adl/\n").expect("seed gitignore");
-    fs::create_dir_all(repo.join("adl/src")).expect("adl src");
-    fs::write(repo.join("adl/src/lib.rs"), "pub fn placeholder() {}\n").expect("write source");
-    assert!(Command::new("git")
-        .args(["add", ".gitignore", "adl/src/lib.rs"])
-        .current_dir(&repo)
-        .status()
-        .expect("git add")
-        .success());
-    assert!(Command::new("git")
-        .args(["commit", "-q", "-m", "init"])
-        .current_dir(&repo)
-        .status()
-        .expect("git commit")
-        .success());
-    assert!(Command::new("git")
-        .args(["branch", "-M", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git branch")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "init",
-            "--bare",
-            "-q",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git init bare")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "remote",
-            "set-url",
-            "origin",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git remote set-url")
-        .success());
-    assert!(Command::new("git")
-        .args(["push", "-q", "-u", "origin", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git push")
-        .success());
-    assert!(Command::new("git")
-        .args(["checkout", "-q", "-b", "codex/1153-rust-finish-test",])
-        .current_dir(&repo)
-        .status()
-        .expect("git checkout")
-        .success());
-
-    let issue_ref =
-        IssueRef::new(1153, "v0.86".to_string(), "rust-finish-test".to_string()).expect("ref");
-    let bundle_dir = issue_ref.task_bundle_dir_path(&repo);
-    fs::create_dir_all(&bundle_dir).expect("bundle dir");
-    let stp = issue_ref.task_bundle_stp_path(&repo);
-    let input = issue_ref.task_bundle_input_path(&repo);
-    let output = issue_ref.task_bundle_output_path(&repo);
-    let plan = issue_ref.task_bundle_plan_path(&repo);
-    let review_policy = issue_ref.task_bundle_review_policy_path(&repo);
-    write_authored_issue_prompt(&repo, &issue_ref, "[v0.86][tools] Rust finish test");
-    fs::copy(issue_ref.issue_prompt_path(&repo), &stp).expect("seed stp");
-    write_authored_sip(
-        &input,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1153-rust-finish-test",
-        &issue_ref.issue_prompt_path(&repo),
-        &repo,
-    );
-    write_authored_spp(
-        &plan,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1153-rust-finish-test",
-        &repo,
-    );
-    write_authored_srp(
-        &review_policy,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1153-rust-finish-test",
-        &repo,
-    );
-    write_completed_sor_fixture(&output, "codex/1153-rust-finish-test");
-
-    fs::write(
-        repo.join("adl/src/lib.rs"),
-        "pub fn placeholder() {}\npub fn changed() {}\n",
-    )
-    .expect("write change");
-
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_log = temp.join("gh.log");
-    let janitor_log = temp.join("janitor.log");
-    let closeout_log = temp.join("closeout.log");
-    let gh_path = bin_dir.join("gh");
-    let janitor_path = bin_dir.join("janitor");
-    let closeout_path = bin_dir.join("closeout");
-    write_executable(
-            &gh_path,
-            &format!(
-                "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" >> '{}'\nif [ \"$1 $2 $3\" = 'repo view --json' ]; then\n  printf 'danielbaustin/agent-design-language\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr list' ]; then\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr create' ]; then\n  printf 'https://github.com/danielbaustin/agent-design-language/pull/1159\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    printf '1153\\n'\n  else\n    printf 'Closes #1153\\n'\n  fi\n  exit 0\nfi\nexit 1\n",
-                gh_log.display()
-            ),
-        );
-    write_executable(
-        &janitor_path,
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" >> '{}'\n",
-            janitor_log.display()
-        ),
-    );
-    write_executable(
-        &closeout_path,
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" >> '{}'\n",
-            closeout_log.display()
-        ),
-    );
-
-    let old_path = env::var("PATH").unwrap_or_default();
-    let old_janitor_cmd = env::var("ADL_PR_JANITOR_CMD").ok();
-    let old_janitor_disable = env::var("ADL_PR_JANITOR_DISABLE").ok();
-    let old_closeout_cmd = env::var("ADL_POST_MERGE_CLOSEOUT_CMD").ok();
-    let old_closeout_disable = env::var("ADL_POST_MERGE_CLOSEOUT_DISABLE").ok();
-    let prev_dir = env::current_dir().expect("cwd");
-    unsafe {
-        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
-        env::set_var("ADL_PR_JANITOR_DISABLE", "0");
-        env::set_var("ADL_PR_JANITOR_CMD", &janitor_path);
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", "0");
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", &closeout_path);
-    }
-    env::set_current_dir(&repo).expect("chdir");
-
-    let result = real_pr(&[
-        "finish".to_string(),
-        "1153".to_string(),
-        "--title".to_string(),
-        "[v0.86][tools] Rust finish test".to_string(),
-        "--paths".to_string(),
-        "adl".to_string(),
-        "--input".to_string(),
-        path_relative_to_repo(&repo, &input),
-        "--output".to_string(),
-        path_relative_to_repo(&repo, &output),
-        "--no-checks".to_string(),
-        "--no-open".to_string(),
-    ]);
-
-    env::set_current_dir(prev_dir).expect("restore cwd");
-    unsafe {
-        env::set_var("PATH", old_path);
-        if let Some(value) = old_janitor_cmd {
-            env::set_var("ADL_PR_JANITOR_CMD", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_CMD");
-        }
-        if let Some(value) = old_janitor_disable {
-            env::set_var("ADL_PR_JANITOR_DISABLE", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_DISABLE");
-        }
-        if let Some(value) = old_closeout_cmd {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        }
-        if let Some(value) = old_closeout_disable {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_DISABLE");
-        }
-    }
-    result.expect("real_pr finish");
-
-    let head_subject = run_capture(
-        "git",
-        &[
-            "-C",
-            path_str(&repo).expect("repo"),
-            "log",
-            "-1",
-            "--format=%s",
-        ],
-    )
-    .expect("head subject");
-    assert!(head_subject.contains("[v0.86][tools] Rust finish test (Closes #1153)"));
-    let head_files = run_capture(
-        "git",
-        &[
-            "-C",
-            path_str(&repo).expect("repo"),
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    )
-    .expect("head files");
-    assert!(!head_files.contains(".adl/v0.86/bodies/issue-1153-rust-finish-test.md"));
-    assert!(!head_files.contains(".adl/v0.86/tasks/issue-1153__rust-finish-test/stp.md"));
-    assert!(!head_files.contains(".adl/v0.86/tasks/issue-1153__rust-finish-test/sip.md"));
-    assert!(!head_files.contains(".adl/v0.86/tasks/issue-1153__rust-finish-test/sor.md"));
-    assert!(Command::new("git")
-        .args([
-            "--git-dir",
-            path_str(&origin).expect("origin"),
-            "rev-parse",
-            "--verify",
-            "refs/heads/codex/1153-rust-finish-test",
-        ])
-        .status()
-        .expect("verify pushed branch")
-        .success());
-    let gh_calls = fs::read_to_string(&gh_log).expect("read gh log");
-    let janitor_calls = fs::read_to_string(&janitor_log).expect("read janitor log");
-    let closeout_calls = fs::read_to_string(&closeout_log).expect("read closeout log");
-    assert!(gh_calls.contains("pr create"));
-    assert!(gh_calls.contains("pr view -R danielbaustin/agent-design-language https://github.com/danielbaustin/agent-design-language/pull/1159 --json closingIssuesReferences --jq .closingIssuesReferences[]?.number"));
-    assert!(janitor_calls.contains("--issue 1153"));
-    assert!(janitor_calls.contains("--branch codex/1153-rust-finish-test"));
-    assert!(janitor_calls
-        .contains("--pr-url https://github.com/danielbaustin/agent-design-language/pull/1159"));
-    assert!(janitor_calls.contains("--expected-pr-state draft"));
-    assert!(closeout_calls.contains("--issue 1153"));
-    assert!(closeout_calls.contains("--branch codex/1153-rust-finish-test"));
-    assert!(closeout_calls
-        .contains("--pr-url https://github.com/danielbaustin/agent-design-language/pull/1159"));
-}
-
-#[test]
-fn real_pr_finish_fails_when_pr_janitor_auto_attach_fails() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-janitor-fail");
-    let origin = temp.join("origin.git");
-    let repo = temp.join("repo");
-    fs::create_dir_all(&repo).expect("repo dir");
-    copy_bootstrap_support_files(&repo);
-    init_git_repo(&repo);
-    assert!(Command::new("git")
-        .args(["config", "user.name", "Test User"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    assert!(Command::new("git")
-        .args(["config", "user.email", "test@example.com"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    fs::write(repo.join(".gitignore"), ".adl/\n").expect("seed gitignore");
-    fs::create_dir_all(repo.join("adl/src")).expect("adl src");
-    fs::write(repo.join("adl/src/lib.rs"), "pub fn placeholder() {}\n").expect("write source");
-    assert!(Command::new("git")
-        .args(["add", ".gitignore", "adl/src/lib.rs"])
-        .current_dir(&repo)
-        .status()
-        .expect("git add")
-        .success());
-    assert!(Command::new("git")
-        .args(["commit", "-q", "-m", "init"])
-        .current_dir(&repo)
-        .status()
-        .expect("git commit")
-        .success());
-    assert!(Command::new("git")
-        .args(["branch", "-M", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git branch")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "init",
-            "--bare",
-            "-q",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git init bare")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "remote",
-            "set-url",
-            "origin",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git remote set-url")
-        .success());
-    assert!(Command::new("git")
-        .args(["push", "-q", "-u", "origin", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git push")
-        .success());
-    assert!(Command::new("git")
-        .args(["checkout", "-q", "-b", "codex/1154-rust-finish-test",])
-        .current_dir(&repo)
-        .status()
-        .expect("git checkout")
-        .success());
-
-    let issue_ref =
-        IssueRef::new(1154, "v0.86".to_string(), "rust-finish-test".to_string()).expect("ref");
-    let bundle_dir = issue_ref.task_bundle_dir_path(&repo);
-    fs::create_dir_all(&bundle_dir).expect("bundle dir");
-    let stp = issue_ref.task_bundle_stp_path(&repo);
-    let input = issue_ref.task_bundle_input_path(&repo);
-    let output = issue_ref.task_bundle_output_path(&repo);
-    let plan = issue_ref.task_bundle_plan_path(&repo);
-    let review_policy = issue_ref.task_bundle_review_policy_path(&repo);
-    write_authored_issue_prompt(&repo, &issue_ref, "[v0.86][tools] Rust finish test");
-    fs::copy(issue_ref.issue_prompt_path(&repo), &stp).expect("seed stp");
-    write_authored_sip(
-        &input,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1154-rust-finish-test",
-        &issue_ref.issue_prompt_path(&repo),
-        &repo,
-    );
-    write_authored_spp(
-        &plan,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1154-rust-finish-test",
-        &repo,
-    );
-    write_authored_srp(
-        &review_policy,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1154-rust-finish-test",
-        &repo,
-    );
-    write_completed_sor_fixture(&output, "codex/1154-rust-finish-test");
-    fs::write(
-        repo.join("adl/src/lib.rs"),
-        "pub fn placeholder() {}\npub fn changed() {}\n",
-    )
-    .expect("write change");
-
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_path = bin_dir.join("gh");
-    let janitor_path = bin_dir.join("janitor");
-    let closeout_path = bin_dir.join("closeout");
-    write_executable(
-        &gh_path,
-        "#!/usr/bin/env bash\nset -euo pipefail\nif [ \"$1 $2 $3\" = 'repo view --json' ]; then\n  printf 'danielbaustin/agent-design-language\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr list' ]; then\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr create' ]; then\n  printf 'https://github.com/danielbaustin/agent-design-language/pull/1160\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    printf '1154\\n'\n  else\n    printf 'Closes #1154\\n'\n  fi\n  exit 0\nfi\nexit 1\n",
-    );
-    write_executable(
-        &janitor_path,
-        "#!/usr/bin/env bash\nset -euo pipefail\necho 'janitor attach failed' >&2\nexit 9\n",
-    );
-    write_executable(
-        &closeout_path,
-        "#!/usr/bin/env bash\nset -euo pipefail\nexit 0\n",
-    );
-
-    let old_path = env::var("PATH").unwrap_or_default();
-    let old_janitor_cmd = env::var("ADL_PR_JANITOR_CMD").ok();
-    let old_janitor_disable = env::var("ADL_PR_JANITOR_DISABLE").ok();
-    let old_closeout_cmd = env::var("ADL_POST_MERGE_CLOSEOUT_CMD").ok();
-    let old_closeout_disable = env::var("ADL_POST_MERGE_CLOSEOUT_DISABLE").ok();
-    let prev_dir = env::current_dir().expect("cwd");
-    unsafe {
-        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
-        env::set_var("ADL_PR_JANITOR_DISABLE", "0");
-        env::set_var("ADL_PR_JANITOR_CMD", &janitor_path);
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", "0");
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", &closeout_path);
-    }
-    env::set_current_dir(&repo).expect("chdir");
-    let result = real_pr(&[
-        "finish".to_string(),
-        "1154".to_string(),
-        "--title".to_string(),
-        "[v0.86][tools] Rust finish test".to_string(),
-        "--paths".to_string(),
-        "adl".to_string(),
-        "--input".to_string(),
-        path_relative_to_repo(&repo, &input),
-        "--output".to_string(),
-        path_relative_to_repo(&repo, &output),
-        "--no-checks".to_string(),
-        "--no-open".to_string(),
-        "--no-close".to_string(),
-    ]);
-    env::set_current_dir(prev_dir).expect("restore cwd");
-    unsafe {
-        env::set_var("PATH", old_path);
-        if let Some(value) = old_janitor_cmd {
-            env::set_var("ADL_PR_JANITOR_CMD", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_CMD");
-        }
-        if let Some(value) = old_janitor_disable {
-            env::set_var("ADL_PR_JANITOR_DISABLE", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_DISABLE");
-        }
-        if let Some(value) = old_closeout_cmd {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        }
-        if let Some(value) = old_closeout_disable {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_DISABLE");
-        }
-    }
-
-    let err = result.expect_err("finish should fail when janitor attach fails");
-    assert!(err
-        .to_string()
-        .contains("finish: PR janitor auto-attach failed"));
-}
-
-#[test]
-fn real_pr_finish_fails_when_post_merge_closeout_auto_attach_fails() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-closeout-fail");
-    let origin = temp.join("origin.git");
-    let repo = temp.join("repo");
-    fs::create_dir_all(&repo).expect("repo dir");
-    copy_bootstrap_support_files(&repo);
-    init_git_repo(&repo);
-    assert!(Command::new("git")
-        .args(["config", "user.name", "Test User"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    assert!(Command::new("git")
-        .args(["config", "user.email", "test@example.com"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    fs::create_dir_all(repo.join("adl/src")).expect("adl src");
-    fs::write(repo.join("adl/src/lib.rs"), "pub fn placeholder() {}\n").expect("write source");
-    assert!(Command::new("git")
-        .args(["add", "-A"])
-        .current_dir(&repo)
-        .status()
-        .expect("git add")
-        .success());
-    assert!(Command::new("git")
-        .args(["commit", "-q", "-m", "init"])
-        .current_dir(&repo)
-        .status()
-        .expect("git commit")
-        .success());
-    assert!(Command::new("git")
-        .args(["branch", "-M", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git branch")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "init",
-            "--bare",
-            "-q",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git init bare")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "remote",
-            "set-url",
-            "origin",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git remote set-url")
-        .success());
-    assert!(Command::new("git")
-        .args(["push", "-q", "-u", "origin", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git push")
-        .success());
-    assert!(Command::new("git")
-        .args(["checkout", "-q", "-b", "codex/1161-rust-finish-test"])
-        .current_dir(&repo)
-        .status()
-        .expect("git checkout")
-        .success());
-
-    let issue_ref =
-        IssueRef::new(1161, "v0.86".to_string(), "rust-finish-test".to_string()).expect("ref");
-    let bundle_dir = issue_ref.task_bundle_dir_path(&repo);
-    fs::create_dir_all(&bundle_dir).expect("bundle dir");
-    let stp = issue_ref.task_bundle_stp_path(&repo);
-    let input = issue_ref.task_bundle_input_path(&repo);
-    let output = issue_ref.task_bundle_output_path(&repo);
-    let plan = issue_ref.task_bundle_plan_path(&repo);
-    let review_policy = issue_ref.task_bundle_review_policy_path(&repo);
-    write_authored_issue_prompt(&repo, &issue_ref, "[v0.86][tools] Rust finish test");
-    fs::copy(issue_ref.issue_prompt_path(&repo), &stp).expect("seed stp");
-    write_authored_sip(
-        &input,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1161-rust-finish-test",
-        &issue_ref.issue_prompt_path(&repo),
-        &repo,
-    );
-    write_authored_spp(
-        &plan,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1161-rust-finish-test",
-        &repo,
-    );
-    write_authored_srp(
-        &review_policy,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test",
-        "codex/1161-rust-finish-test",
-        &repo,
-    );
-    write_completed_sor_fixture(&output, "codex/1161-rust-finish-test");
-    fs::write(
-        repo.join("adl/src/lib.rs"),
-        "pub fn placeholder() {}\npub fn changed() {}\n",
-    )
-    .expect("write change");
-
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_path = bin_dir.join("gh");
-    let janitor_path = bin_dir.join("janitor");
-    let closeout_path = bin_dir.join("closeout");
-    write_executable(
-        &gh_path,
-        "#!/usr/bin/env bash\nset -euo pipefail\nif [ \"$1 $2 $3\" = 'repo view --json' ]; then\n  printf 'danielbaustin/agent-design-language\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr list' ]; then\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr create' ]; then\n  printf 'https://github.com/danielbaustin/agent-design-language/pull/1161\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    printf '1161\\n'\n  else\n    printf 'Closes #1161\\n'\n  fi\n  exit 0\nfi\nexit 1\n",
-    );
-    write_executable(
-        &janitor_path,
-        "#!/usr/bin/env bash\nset -euo pipefail\nexit 0\n",
-    );
-    write_executable(
-        &closeout_path,
-        "#!/usr/bin/env bash\nset -euo pipefail\necho 'closeout attach failed' >&2\nexit 9\n",
-    );
-
-    let old_path = env::var("PATH").unwrap_or_default();
-    let old_janitor_cmd = env::var("ADL_PR_JANITOR_CMD").ok();
-    let old_janitor_disable = env::var("ADL_PR_JANITOR_DISABLE").ok();
-    let old_closeout_cmd = env::var("ADL_POST_MERGE_CLOSEOUT_CMD").ok();
-    let old_closeout_disable = env::var("ADL_POST_MERGE_CLOSEOUT_DISABLE").ok();
-    let prev_dir = env::current_dir().expect("cwd");
-    unsafe {
-        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
-        env::set_var("ADL_PR_JANITOR_DISABLE", "0");
-        env::set_var("ADL_PR_JANITOR_CMD", &janitor_path);
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", "0");
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", &closeout_path);
-    }
-    env::set_current_dir(&repo).expect("chdir");
-    let result = real_pr(&[
-        "finish".to_string(),
-        "1161".to_string(),
-        "--title".to_string(),
-        "[v0.86][tools] Rust finish test".to_string(),
-        "--paths".to_string(),
-        "adl".to_string(),
-        "--input".to_string(),
-        path_relative_to_repo(&repo, &input),
-        "--output".to_string(),
-        path_relative_to_repo(&repo, &output),
-        "--no-checks".to_string(),
-        "--no-open".to_string(),
-    ]);
-    env::set_current_dir(prev_dir).expect("restore cwd");
-    unsafe {
-        env::set_var("PATH", old_path);
-        if let Some(value) = old_janitor_cmd {
-            env::set_var("ADL_PR_JANITOR_CMD", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_CMD");
-        }
-        if let Some(value) = old_janitor_disable {
-            env::set_var("ADL_PR_JANITOR_DISABLE", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_DISABLE");
-        }
-        if let Some(value) = old_closeout_cmd {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        }
-        if let Some(value) = old_closeout_disable {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_DISABLE");
-        }
-    }
-
-    let err = result.expect_err("finish should fail when post-merge closeout auto-attach fails");
-    assert!(err
-        .to_string()
-        .contains("finish: post-merge closeout auto-attach failed"));
-}
-
-#[test]
-fn real_pr_finish_updates_existing_pr_and_marks_ready() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-edit");
-    let origin = temp.join("origin.git");
-    let repo = temp.join("repo");
-    fs::create_dir_all(&repo).expect("repo dir");
-    copy_bootstrap_support_files(&repo);
-    init_git_repo(&repo);
-    assert!(Command::new("git")
-        .args(["config", "user.name", "Test User"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    assert!(Command::new("git")
-        .args(["config", "user.email", "test@example.com"])
-        .current_dir(&repo)
-        .status()
-        .expect("git config")
-        .success());
-    fs::create_dir_all(repo.join("adl/src")).expect("adl src");
-    fs::write(repo.join("adl/src/lib.rs"), "pub fn placeholder() {}\n").expect("write source");
-    assert!(Command::new("git")
-        .args(["add", "-A"])
-        .current_dir(&repo)
-        .status()
-        .expect("git add")
-        .success());
-    assert!(Command::new("git")
-        .args(["commit", "-q", "-m", "init"])
-        .current_dir(&repo)
-        .status()
-        .expect("git commit")
-        .success());
-    assert!(Command::new("git")
-        .args(["branch", "-M", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git branch")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "init",
-            "--bare",
-            "-q",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git init bare")
-        .success());
-    assert!(Command::new("git")
-        .args([
-            "remote",
-            "set-url",
-            "origin",
-            path_str(&origin).expect("origin path"),
-        ])
-        .current_dir(&repo)
-        .status()
-        .expect("git remote set-url")
-        .success());
-    assert!(Command::new("git")
-        .args(["push", "-q", "-u", "origin", "main"])
-        .current_dir(&repo)
-        .status()
-        .expect("git push")
-        .success());
-    assert!(Command::new("git")
-        .args(["checkout", "-q", "-b", "codex/1153-rust-finish-test-edit",])
-        .current_dir(&repo)
-        .status()
-        .expect("git checkout")
-        .success());
-
-    let issue_ref = IssueRef::new(
-        1153,
-        "v0.86".to_string(),
-        "rust-finish-test-edit".to_string(),
-    )
-    .expect("ref");
-    let bundle_dir = issue_ref.task_bundle_dir_path(&repo);
-    fs::create_dir_all(&bundle_dir).expect("bundle dir");
-    let stp = issue_ref.task_bundle_stp_path(&repo);
-    let input = issue_ref.task_bundle_input_path(&repo);
-    let output = issue_ref.task_bundle_output_path(&repo);
-    let plan = issue_ref.task_bundle_plan_path(&repo);
-    let review_policy = issue_ref.task_bundle_review_policy_path(&repo);
-    write_authored_issue_prompt(&repo, &issue_ref, "[v0.86][tools] Rust finish test edit");
-    fs::copy(issue_ref.issue_prompt_path(&repo), &stp).expect("seed stp");
-    write_authored_sip(
-        &input,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test edit",
-        "codex/1153-rust-finish-test-edit",
-        &issue_ref.issue_prompt_path(&repo),
-        &repo,
-    );
-    write_authored_spp(
-        &plan,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test edit",
-        "codex/1153-rust-finish-test-edit",
-        &repo,
-    );
-    write_authored_srp(
-        &review_policy,
-        &issue_ref,
-        "[v0.86][tools] Rust finish test edit",
-        "codex/1153-rust-finish-test-edit",
-        &repo,
-    );
-    write_completed_sor_fixture(&output, "codex/1153-rust-finish-test-edit");
-
-    fs::write(
-        repo.join("adl/src/lib.rs"),
-        "pub fn placeholder() -> usize { 1 }\n",
-    )
-    .expect("write change");
-    assert!(Command::new("git")
-        .args(["add", "adl/src/lib.rs"])
-        .current_dir(&repo)
-        .status()
-        .expect("git add")
-        .success());
-    assert!(Command::new("git")
-        .args(["commit", "-q", "-m", "existing branch commit"])
-        .current_dir(&repo)
-        .status()
-        .expect("git commit")
-        .success());
-
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_log = temp.join("gh.log");
-    let gh_path = bin_dir.join("gh");
-    write_executable(
-            &gh_path,
-            &format!(
-                "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" >> '{}'\nif [ \"$1 $2 $3\" = 'repo view --json' ]; then\n  printf 'danielbaustin/agent-design-language\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr list' ]; then\n  printf 'https://github.com/danielbaustin/agent-design-language/pull/1160\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr edit' ]; then\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr ready' ]; then\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    printf '1153\\n'\n  else\n    printf 'Closes #1153\\n'\n  fi\n  exit 0\nfi\nexit 1\n",
-                gh_log.display()
-            ),
-        );
-
-    let old_path = env::var("PATH").unwrap_or_default();
-    let prev_dir = env::current_dir().expect("cwd");
-    unsafe {
-        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
-    }
-    env::set_current_dir(&repo).expect("chdir");
-
-    let result = real_pr(&[
-        "finish".to_string(),
-        "1153".to_string(),
-        "--title".to_string(),
-        "[v0.86][tools] Rust finish test edit".to_string(),
-        "--paths".to_string(),
-        "adl".to_string(),
-        "--input".to_string(),
-        path_relative_to_repo(&repo, &input),
-        "--output".to_string(),
-        path_relative_to_repo(&repo, &output),
-        "--no-checks".to_string(),
-        "--ready".to_string(),
-        "--no-open".to_string(),
-    ]);
-
-    env::set_current_dir(prev_dir).expect("restore cwd");
-    unsafe {
-        env::set_var("PATH", old_path);
-    }
-    result.expect("real_pr finish edit");
-
-    let gh_calls = fs::read_to_string(&gh_log).expect("read gh log");
-    assert!(gh_calls.contains("pr edit -R danielbaustin/agent-design-language https://github.com/danielbaustin/agent-design-language/pull/1160 --title [v0.86][tools] Rust finish test edit --body-file"));
-    assert!(gh_calls.contains("pr ready -R danielbaustin/agent-design-language https://github.com/danielbaustin/agent-design-language/pull/1160"));
-}
-
-#[test]
-fn finish_helper_paths_cover_pr_lookup_and_closing_linkage() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-gh-helpers");
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_log = temp.join("gh.log");
-    let gh_path = bin_dir.join("gh");
-    write_executable(
-            &gh_path,
-            &format!(
-                "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" >> '{}'\nif [ \"$1 $2\" = 'pr list' ]; then\n  printf 'https://github.com/danielbaustin/agent-design-language/pull/1159\\n'\n  exit 0\nfi\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    printf '1153\\n'\n  else\n    printf 'Closes #1153\\n'\n  fi\n  exit 0\nfi\nexit 1\n",
-                gh_log.display()
-            ),
-        );
-    let old_path = env::var("PATH").unwrap_or_default();
-    unsafe {
-        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
-    }
-
-    let pr = current_pr_url(
-        "danielbaustin/agent-design-language",
-        "codex/1153-rust-finish-test",
-    )
-    .expect("pr url");
-    assert_eq!(
-        pr.as_deref(),
-        Some("https://github.com/danielbaustin/agent-design-language/pull/1159")
-    );
-    assert!(pr_has_closing_linkage(
-        "danielbaustin/agent-design-language",
-        "https://github.com/danielbaustin/agent-design-language/pull/1159",
-        1153
-    )
-    .expect("closing linkage"));
-    ensure_pr_closing_linkage(
-        "danielbaustin/agent-design-language",
-        "https://github.com/danielbaustin/agent-design-language/pull/1159",
-        1153,
-        false,
-    )
-    .expect("ensure linkage");
-    ensure_pr_closing_linkage(
-        "danielbaustin/agent-design-language",
-        "https://github.com/danielbaustin/agent-design-language/pull/1159",
-        1153,
-        true,
-    )
-    .expect("no close skips");
-
-    unsafe {
-        env::set_var("PATH", old_path);
-    }
-    let gh_calls = fs::read_to_string(&gh_log).expect("gh log");
-    assert!(gh_calls.contains("pr list -R danielbaustin/agent-design-language --head codex/1153-rust-finish-test --state open --json url --jq .[0].url"));
-}
-
-#[test]
-fn ensure_or_repair_pr_closing_linkage_repairs_live_pr_body() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-repair-linkage");
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_log = temp.join("gh.log");
-    let state_body = temp.join("pr_body.txt");
-    fs::write(&state_body, "Refs #1153\n").expect("seed body");
-    let gh_path = bin_dir.join("gh");
-    write_executable(
-        &gh_path,
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" >> '{}'\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    if grep -q 'Closes #1153' '{}'; then\n      printf '1153\\n'\n    fi\n    exit 0\n  fi\n  if printf '%s ' \"$@\" | grep -q ' --json body '; then\n    cat '{}'\n    exit 0\n  fi\nfi\nif [ \"$1 $2\" = 'pr edit' ]; then\n  body_file=''\n  while [ $# -gt 0 ]; do\n    if [ \"$1\" = '--body-file' ]; then\n      body_file=\"$2\"\n      shift 2\n    else\n      shift\n    fi\n  done\n  cp \"$body_file\" '{}'\n  exit 0\nfi\nexit 1\n",
-            gh_log.display(),
-            state_body.display(),
-            state_body.display(),
-            state_body.display()
-        ),
-    );
-
-    let body_file = temp.join("desired.md");
-    fs::write(&body_file, "Closes #1153\n\n## Summary\nrepaired\n").expect("desired body");
-
-    let old_path = env::var("PATH").unwrap_or_default();
-    let old_entries = env::split_paths(&old_path).collect::<Vec<_>>();
-    let mut new_entries = vec![bin_dir.clone()];
-    new_entries.extend(old_entries);
-    unsafe {
-        env::set_var("PATH", env::join_paths(new_entries).expect("join PATH"));
-    }
-
-    ensure_or_repair_pr_closing_linkage(
-        "danielbaustin/agent-design-language",
-        "https://github.com/danielbaustin/agent-design-language/pull/1159",
-        1153,
-        false,
-        &body_file,
-    )
-    .expect("repair should succeed");
-
-    unsafe {
-        env::set_var("PATH", old_path);
-    }
-
-    let repaired = fs::read_to_string(&state_body).expect("read repaired body");
-    assert!(repaired.contains("Closes #1153"));
-    let gh_calls = fs::read_to_string(&gh_log).expect("read gh log");
-    assert!(gh_calls.contains("pr edit -R danielbaustin/agent-design-language https://github.com/danielbaustin/agent-design-language/pull/1159 --body-file"));
-}
-
-#[test]
-fn ensure_pr_closing_linkage_errors_when_pr_body_has_no_issue_reference() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-missing-linkage");
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_path = bin_dir.join("gh");
-    write_executable(
-        &gh_path,
-        "#!/usr/bin/env bash\nset -euo pipefail\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    exit 0\n  fi\n  if printf '%s ' \"$@\" | grep -q ' --json body '; then\n    printf 'Refs #9999\\n'\n    exit 0\n  fi\nfi\nexit 1\n",
-    );
-
-    let old_path = env::var("PATH").unwrap_or_default();
-    unsafe {
-        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
-    }
-
-    let err = ensure_pr_closing_linkage(
-        "danielbaustin/agent-design-language",
-        "https://github.com/danielbaustin/agent-design-language/pull/1159",
-        1153,
-        false,
-    )
-    .expect_err("missing linkage should fail");
-
-    unsafe {
-        env::set_var("PATH", old_path);
-    }
-
-    assert!(err
-        .to_string()
-        .contains("missing closing linkage to issue #1153"));
-}
-
-#[test]
-fn ensure_or_repair_pr_closing_linkage_is_noop_when_no_close_is_set() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-no-close");
-    let body_file = temp.join("desired.md");
-    fs::write(&body_file, "Refs #1153\n").expect("body");
-
-    let repaired = ensure_or_repair_pr_closing_linkage(
-        "danielbaustin/agent-design-language",
-        "https://github.com/danielbaustin/agent-design-language/pull/1159",
-        1153,
-        true,
-        &body_file,
-    )
-    .expect("no-close should skip linkage repair");
-
-    assert!(!repaired);
-}
-
-#[test]
-fn ensure_or_repair_pr_closing_linkage_is_noop_when_linkage_already_exists() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-finish-linkage-already-present");
-    let bin_dir = temp.join("bin");
-    fs::create_dir_all(&bin_dir).expect("bin dir");
-    let gh_log = temp.join("gh.log");
-    write_executable(
-        &bin_dir.join("gh"),
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" >> '{}'\nif [ \"$1 $2\" = 'pr view' ]; then\n  if printf '%s ' \"$@\" | grep -q 'closingIssuesReferences'; then\n    printf '1153\\n'\n    exit 0\n  fi\n  if printf '%s ' \"$@\" | grep -q ' --json body '; then\n    printf 'Closes #1153\\n'\n    exit 0\n  fi\nfi\nif [ \"$1 $2\" = 'pr edit' ]; then\n  exit 99\nfi\nexit 1\n",
-            gh_log.display()
-        ),
-    );
-    let body_file = temp.join("desired.md");
-    fs::write(&body_file, "Closes #1153\n").expect("body");
-
-    let old_path = env::var("PATH").unwrap_or_default();
-    unsafe {
-        env::set_var("PATH", format!("{}:{}", bin_dir.display(), old_path));
-    }
-
-    let repaired = ensure_or_repair_pr_closing_linkage(
-        "danielbaustin/agent-design-language",
-        "https://github.com/danielbaustin/agent-design-language/pull/1159",
-        1153,
-        false,
-        &body_file,
-    )
-    .expect("existing linkage should not require repair");
-
-    unsafe {
-        env::set_var("PATH", old_path);
-    }
-
-    assert!(!repaired);
-    let gh_log = fs::read_to_string(&gh_log).expect("gh log");
-    assert!(!gh_log.contains("pr edit -R"));
-}
-
-#[test]
-fn attach_pr_janitor_reports_failure_output() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-janitor-failure");
-    let repo = temp.join("repo");
-    let tools_dir = repo.join("adl/tools");
-    fs::create_dir_all(&tools_dir).expect("tools dir");
-    write_executable(
-        &tools_dir.join("attach_pr_janitor.sh"),
-        "#!/usr/bin/env bash\nset -euo pipefail\necho 'janitor stdout'\necho 'janitor stderr' >&2\nexit 17\n",
-    );
-
-    let old_disable = env::var("ADL_PR_JANITOR_DISABLE").ok();
-    let old_cmd = env::var("ADL_PR_JANITOR_CMD").ok();
-    unsafe {
-        env::remove_var("ADL_PR_JANITOR_CMD");
-        env::set_var("ADL_PR_JANITOR_DISABLE", "0");
-    }
-
-    let err = attach_pr_janitor(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-        "draft",
-    )
-    .expect_err("failing janitor helper should bubble up");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_PR_JANITOR_DISABLE", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_PR_JANITOR_CMD", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_CMD");
-        }
-    }
-
-    let message = err.to_string();
-    assert!(message.contains("PR janitor auto-attach failed"));
-    assert!(message.contains("janitor stderr"));
-    assert!(message.contains("stdout: janitor stdout"));
-}
-
-#[test]
-fn attach_pr_janitor_returns_early_when_disabled() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-janitor-disabled");
-    let repo = temp.join("repo");
-
-    let old_disable = env::var("ADL_PR_JANITOR_DISABLE").ok();
-    let old_cmd = env::var("ADL_PR_JANITOR_CMD").ok();
-    unsafe {
-        env::set_var("ADL_PR_JANITOR_DISABLE", "1");
-        env::remove_var("ADL_PR_JANITOR_CMD");
-    }
-
-    attach_pr_janitor(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-        "draft",
-    )
-    .expect("disabled janitor helper should be skipped");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_PR_JANITOR_DISABLE", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_PR_JANITOR_CMD", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_CMD");
-        }
-    }
-}
-
-#[test]
-fn attach_pr_janitor_invokes_helper_successfully() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-janitor-success");
-    let repo = temp.join("repo");
-    let tools_dir = repo.join("adl/tools");
-    let argv_log = temp.join("janitor-args.log");
-    fs::create_dir_all(&tools_dir).expect("tools dir");
-    write_executable(
-        &tools_dir.join("attach_pr_janitor.sh"),
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" > '{}'\n",
-            argv_log.display()
-        ),
-    );
-
-    let old_disable = env::var("ADL_PR_JANITOR_DISABLE").ok();
-    let old_cmd = env::var("ADL_PR_JANITOR_CMD").ok();
-    unsafe {
-        env::set_var("ADL_PR_JANITOR_DISABLE", "0");
-        env::remove_var("ADL_PR_JANITOR_CMD");
-    }
-
-    attach_pr_janitor(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-        "draft",
-    )
-    .expect("janitor helper should succeed");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_PR_JANITOR_DISABLE", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_PR_JANITOR_CMD", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_CMD");
-        }
-    }
-
-    let argv = fs::read_to_string(&argv_log).expect("janitor args");
-    assert!(argv.contains("--repo owner/repo"));
-    assert!(argv.contains("--issue 1153"));
-    assert!(argv.contains("--branch codex/1153-rust-finish-test"));
-    assert!(argv.contains("--pr-url https://github.com/owner/repo/pull/1159"));
-    assert!(argv.contains("--expected-pr-state draft"));
-}
-
-#[test]
-fn attach_pr_janitor_falls_back_when_command_override_is_blank() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-janitor-blank-override");
-    let repo = temp.join("repo");
-    let tools_dir = repo.join("adl/tools");
-    let argv_log = temp.join("janitor-args.log");
-    fs::create_dir_all(&tools_dir).expect("tools dir");
-    write_executable(
-        &tools_dir.join("attach_pr_janitor.sh"),
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" > '{}'\n",
-            argv_log.display()
-        ),
-    );
-
-    let old_disable = env::var("ADL_PR_JANITOR_DISABLE").ok();
-    let old_cmd = env::var("ADL_PR_JANITOR_CMD").ok();
-    unsafe {
-        env::set_var("ADL_PR_JANITOR_DISABLE", "0");
-        env::set_var("ADL_PR_JANITOR_CMD", "   ");
-    }
-
-    attach_pr_janitor(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-        "draft",
-    )
-    .expect("blank command override should fall back to helper path");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_PR_JANITOR_DISABLE", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_PR_JANITOR_CMD", value);
-        } else {
-            env::remove_var("ADL_PR_JANITOR_CMD");
-        }
-    }
-
-    let argv = fs::read_to_string(&argv_log).expect("janitor args");
-    assert!(argv.contains("--repo owner/repo"));
-}
-
-#[test]
-fn attach_post_merge_closeout_reports_failure_output() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-closeout-failure");
-    let repo = temp.join("repo");
-    let tools_dir = repo.join("adl/tools");
-    fs::create_dir_all(&tools_dir).expect("tools dir");
-    write_executable(
-        &tools_dir.join("attach_post_merge_closeout.sh"),
-        "#!/usr/bin/env bash\nset -euo pipefail\necho 'closeout stdout'\necho 'closeout stderr' >&2\nexit 18\n",
-    );
-
-    let old_disable = env::var("ADL_POST_MERGE_CLOSEOUT_DISABLE").ok();
-    let old_cmd = env::var("ADL_POST_MERGE_CLOSEOUT_CMD").ok();
-    unsafe {
-        env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", "0");
-    }
-
-    let err = attach_post_merge_closeout(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-    )
-    .expect_err("failing closeout helper should bubble up");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        }
-    }
-
-    let message = err.to_string();
-    assert!(message.contains("post-merge closeout auto-attach failed"));
-    assert!(message.contains("closeout stderr"));
-    assert!(message.contains("stdout: closeout stdout"));
-}
-
-#[test]
-fn attach_post_merge_closeout_returns_early_when_disabled() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-closeout-disabled");
-    let repo = temp.join("repo");
-
-    let old_disable = env::var("ADL_POST_MERGE_CLOSEOUT_DISABLE").ok();
-    let old_cmd = env::var("ADL_POST_MERGE_CLOSEOUT_CMD").ok();
-    unsafe {
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", "1");
-        env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-    }
-
-    attach_post_merge_closeout(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-    )
-    .expect("disabled closeout helper should be skipped");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        }
-    }
-}
-
-#[test]
-fn attach_post_merge_closeout_invokes_helper_successfully() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-closeout-success");
-    let repo = temp.join("repo");
-    let tools_dir = repo.join("adl/tools");
-    let argv_log = temp.join("closeout-args.log");
-    fs::create_dir_all(&tools_dir).expect("tools dir");
-    write_executable(
-        &tools_dir.join("attach_post_merge_closeout.sh"),
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" > '{}'\n",
-            argv_log.display()
-        ),
-    );
-
-    let old_disable = env::var("ADL_POST_MERGE_CLOSEOUT_DISABLE").ok();
-    let old_cmd = env::var("ADL_POST_MERGE_CLOSEOUT_CMD").ok();
-    unsafe {
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", "0");
-        env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-    }
-
-    attach_post_merge_closeout(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-    )
-    .expect("closeout helper should succeed");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        }
-    }
-
-    let argv = fs::read_to_string(&argv_log).expect("closeout args");
-    assert!(argv.contains("--repo owner/repo"));
-    assert!(argv.contains("--issue 1153"));
-    assert!(argv.contains("--branch codex/1153-rust-finish-test"));
-    assert!(argv.contains("--pr-url https://github.com/owner/repo/pull/1159"));
-}
-
-#[test]
-fn attach_post_merge_closeout_falls_back_when_command_override_is_blank() {
-    let _guard = env_lock();
-    let temp = unique_temp_dir("adl-pr-attach-closeout-blank-override");
-    let repo = temp.join("repo");
-    let tools_dir = repo.join("adl/tools");
-    let argv_log = temp.join("closeout-args.log");
-    fs::create_dir_all(&tools_dir).expect("tools dir");
-    write_executable(
-        &tools_dir.join("attach_post_merge_closeout.sh"),
-        &format!(
-            "#!/usr/bin/env bash\nset -euo pipefail\nprintf '%s\\n' \"$*\" > '{}'\n",
-            argv_log.display()
-        ),
-    );
-
-    let old_disable = env::var("ADL_POST_MERGE_CLOSEOUT_DISABLE").ok();
-    let old_cmd = env::var("ADL_POST_MERGE_CLOSEOUT_CMD").ok();
-    unsafe {
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", "0");
-        env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", "   ");
-    }
-
-    attach_post_merge_closeout(
-        &repo,
-        "owner/repo",
-        1153,
-        "codex/1153-rust-finish-test",
-        "https://github.com/owner/repo/pull/1159",
-    )
-    .expect("blank command override should fall back to helper path");
-
-    unsafe {
-        if let Some(value) = old_disable {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_DISABLE", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_DISABLE");
-        }
-        if let Some(value) = old_cmd {
-            env::set_var("ADL_POST_MERGE_CLOSEOUT_CMD", value);
-        } else {
-            env::remove_var("ADL_POST_MERGE_CLOSEOUT_CMD");
-        }
-    }
-
-    let argv = fs::read_to_string(&argv_log).expect("closeout args");
-    assert!(argv.contains("--repo owner/repo"));
-}
+mod closeout;
+mod flow;
+mod janitor;
+mod linkage;
